whack-a-mole!
How to Play:
Start Game: Click the "Start Game" button. Moles will start popping up randomly from the holes.

Whack Moles: Click on a mole as soon as it appears to "whack" it and earn points.

Lives: If a mole appears and disappears without being whacked, you lose a life. The game ends if you run out of lives or time.